Stripping BOM to make it Windows compatible
Be default, most Windows text editors saves files in UTF-8 encoding with a signature (BOM). The async `fs.readFile` method in Node.js doesn't strip the BOM automatically, which makes most files invalid. In the case of TV4, all `.json` files are unreadable on Windows because of this issue.

The official workaround is to strip the BOM like I did in this PR.